@@ -1,8 +1,10 @@
 package cmd
 
 import (
+	"bytes"
 	"encoding/json"
 	"fmt"
+	"io"
 	"net/http"
 	"sort"
 	"strings"
@@ -25,6 +27,39 @@ func (c *DefaultHTTPClient) Do(req *http.Request) (*http.Response, error) {
 
 var httpClient HTTPClient = &DefaultHTTPClient{}
 
+// logBodyContent safely reads and logs a body, restoring it for later use.
+// Returns the restored body (or nil if input was nil).
+// Truncates very large bodies to avoid flooding logs.
+func logBodyContent(body io.ReadCloser, label string) io.ReadCloser {
+	if body == nil {
+		logDebug(fmt.Sprintf("  -> %s: <nil>", label))
+		return nil
+	}
+
+	bodyBytes, err := io.ReadAll(body)
+	body.Close()
+
+	if err != nil {
+		logDebug(fmt.Sprintf("  -> %s: <error reading: %v>", label, err))
+		return io.NopCloser(bytes.NewReader([]byte{}))
+	}
+
+	if len(bodyBytes) == 0 {
+		logDebug(fmt.Sprintf("  -> %s: <empty>", label))
+		return io.NopCloser(bytes.NewReader(bodyBytes))
+	}
+
+	// Truncate very large bodies for readability
+	const maxLogSize = 1024
+	bodyStr := string(bodyBytes)
+	if len(bodyStr) > maxLogSize {
+		bodyStr = bodyStr[:maxLogSize] + "... (truncated)"
+	}
+
+	logDebug(fmt.Sprintf("  -> %s: %s", label, bodyStr))
+	return io.NopCloser(bytes.NewReader(bodyBytes))
+}
+
 // VerboseHTTPClient wraps another HTTPClient and logs request/response basics and headers.
 type VerboseHTTPClient struct{ Inner HTTPClient }
 
@@ -35,7 +70,10 @@ func (v *VerboseHTTPClient) Do(req *http.Request) (*http.Response, error) {
 	}
 	logDebug(fmt.Sprintf("HTTP %s %s", req.Method, req.URL.String()))
 	logHeaders("request", req.Header)
-	logDebug(fmt.Sprintf("  -> body: %s", req.Body))
+
+	// Log and restore request body
+	req.Body = logBodyContent(req.Body, "request body")
+
 	resp, err := inner.Do(req)
 	if err != nil {
 		logDebug(fmt.Sprintf("  -> error: %v", err))
@@ -43,7 +81,10 @@ func (v *VerboseHTTPClient) Do(req *http.Request) (*http.Response, error) {
 	}
 	logDebug(fmt.Sprintf("  -> %d %s", resp.StatusCode, http.StatusText(resp.StatusCode)))
 	logHeaders("response", resp.Header)
-	logDebug(fmt.Sprintf("  -> body: %s", resp.Body))
+
+	// Log and restore response body
+	resp.Body = logBodyContent(resp.Body, "response body")
+
 	return resp, nil
 }
 
@@ -62,9 +103,9 @@ func logHeaders(kind string, hdr http.Header) {
 	// List of sensitive headers to redact, lower-case for comparison
 	sensitiveHeaders := map[string]struct{}{
 		"authorization": {},
-		"cookie": {},
-		"x-session-id": {},
-		"set-cookie": {},
+		"cookie":        {},
+		"x-session-id":  {},
+		"set-cookie":    {},
 		// Add others as needed
 	}
 	keys := make([]string, 0, len(hdr))
@@ -89,12 +130,12 @@ func logHeaders(kind string, hdr http.Header) {
 // It parses common JSON shapes like {"detail":...}, {"message":...}, {"error":...}.
 func prettyServerError(resp *http.Response, body []byte) string {
 	// Try to parse JSON error envelopes
-    var env struct {
-        Detail    any    `json:"detail"`
-        Message   string `json:"message"`
-        Error     string `json:"error"`
-        RequestID string `json:"request_id"`
-    }
+	var env struct {
+		Detail    any    `json:"detail"`
+		Message   string `json:"message"`
+		Error     string `json:"error"`
+		RequestID string `json:"request_id"`
+	}
 	if json.Unmarshal(body, &env) == nil {
 		switch v := env.Detail.(type) {
 		case string:
@@ -127,13 +168,13 @@ func prettyServerError(resp *http.Response, body []byte) string {
 			return env.Error
 		}
 	}
-    s := strings.TrimSpace(string(body))
+	s := strings.TrimSpace(string(body))
 	if s == "" {
 		return http.StatusText(resp.StatusCode)
 	}
-    // Append request id if present to aid debugging
-    if env.RequestID != "" {
-        return s + " (request_id=" + env.RequestID + ")"
-    }
-    return s
+	// Append request id if present to aid debugging
+	if env.RequestID != "" {
+		return s + " (request_id=" + env.RequestID + ")"
+	}
+	return s
 }
