@@ -1,12 +1,12 @@
-import { useState } from 'react'
-import ModeToggle, { Mode } from '../ModeToggle'
+import ModeToggle from '../ModeToggle'
 import { Button } from '../ui/button'
 import ConfigEditor from '../ConfigEditor/ConfigEditor'
 import { usePackageModal } from '../../contexts/PackageModalContext'
 import Prompts from './GeneratedOutput/Prompts'
+import { useModeWithReset } from '../../hooks/useModeWithReset'
 
 const Prompt = () => {
-  const [mode, setMode] = useState<Mode>('designer')
+  const [mode, setMode] = useModeWithReset('designer')
   const { openPackageModal } = usePackageModal()
 
   return (
@@ -17,7 +17,12 @@ const Prompt = () => {
         </h2>
         <div className="flex items-center gap-3">
           <ModeToggle mode={mode} onToggle={setMode} />
-          <Button variant="outline" size="sm" onClick={openPackageModal}>
+          <Button
+            variant="outline"
+            size="sm"
+            onClick={openPackageModal}
+            disabled
+          >
             Package
           </Button>
         </div>
