@@ -7,9 +7,9 @@ import { Input } from '../ui/input'
 // import { Badge } from '../ui/badge'
 import PageActions from '../common/PageActions'
 import ConfigEditor from '../ConfigEditor/ConfigEditor'
-import { Mode } from '../ModeToggle'
 import { useToast } from '../ui/toast'
 import { Label } from '../ui/label'
+import { useModeWithReset } from '../../hooks/useModeWithReset'
 import {
   DropdownMenu,
   DropdownMenuContent,
@@ -69,7 +69,7 @@ async function encryptAPIKey(apiKey: string, secret: string) {
 
 function ChangeEmbeddingModel() {
   const navigate = useNavigate()
-  const [mode, setMode] = useState<Mode>('designer')
+  const [mode, setMode] = useModeWithReset('designer')
   const { strategyId } = useParams()
   const { toast } = useToast()
 
